Added users data

+++ app/src/main/java/com/example/hackintosh/tennismate/dto/User.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
 public class User implements Serializable {
     private String uuid;
     private String email;
-    private Short age;
+    private int age;
     private String fullName;
     private LevelEnum level;
     private String imageUrl;
@@ -17,7 +17,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(String email, Short age, String fullName, LevelEnum level, String imageUrl) {
+    public User(String email, short age, String fullName, LevelEnum level, String imageUrl) {
         this.email = email;
         this.age = age;
         this.fullName = fullName;
@@ -37,7 +37,7 @@ public class User implements Serializable {
         return age;
     }
 
-    public void setAge(Short age) {
+    public void setAge(int age) {
         this.age = age;
     }
